+++ src/main/java/uk/gov/digital/ho/hocs/cms/domain/ComplaintExtractRecord.java
@@ -11,6 +11,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -26,7 +27,7 @@ public class ComplaintExtractRecord {
     private Long id;
 
     @Column
-    private int caseId;
+    private BigDecimal caseId;
 
     @Column
     private boolean complaintExtracted;
