+++ src/main/java/uk/gov/digital/ho/hocs/cms/domain/repository/DocumentsRepository.java
@@ -5,9 +5,11 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import uk.gov.digital.ho.hocs.cms.domain.DocumentExtractRecord;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface DocumentsRepository extends CrudRepository<DocumentExtractRecord, Long> {
+public interface DocumentsRepository extends CrudRepository<DocumentExtractRecord, BigDecimal> {
 
     @Query(value = "SELECT count(*) FROM documents d WHERE d.case_id = ?1 AND d.document_extracted = 'false'", nativeQuery = true)
-    int findFailedDocumentsForCase(int caseId);
+    BigDecimal findFailedDocumentsForCase(BigDecimal caseId);
 }
