+++ src/main/resources/db/migration/V1_1__CREATE_TABLES.sql
@@ -3,9 +3,9 @@ DROP TABLE IF EXISTS documents cascade;
 CREATE TABLE IF NOT EXISTS documents
 (
     id                         BIGSERIAL PRIMARY KEY,
-    case_id                    INT,
+    case_id                    NUMERIC,
     document_extracted         BOOLEAN,
-    document_id                INT,
+    document_id                NUMERIC,
     failure_reason             TEXT,
     temp_file_name             TEXT
 );
@@ -15,7 +15,7 @@ DROP TABLE IF EXISTS complaints cascade;
 CREATE TABLE IF NOT EXISTS complaints
 (
     id                         BIGSERIAL PRIMARY KEY,
-    case_id                    INT,
+    case_id                    NUMERIC,
     complaint_extracted        BOOLEAN,
     stage                      TEXT
 );
